+++ tracked-items.js
@@ -2,6 +2,14 @@ let trackedItemsContainer;
 let trackedItemsList;
 let listItemsLoader;
 
+const ITEM_STATUS = {
+    WATCHED: "WATCHED",
+    NOT_FOUND: "NOT_FOUND",
+    INCREASED: "INCREASED",
+    DECREASED: "DECREASED",
+    FIXED: "FIXED"
+};
+
 function loadTemplate(elementId) {
     const elementHTML = document.getElementById(elementId).innerHTML;
     return Handlebars.compile(elementHTML);
@@ -15,7 +23,11 @@ function formatDate(timestamp) {
 function onTrackedItems(rawItems) {
     const trackedItems = rawItems.map(item => ({
         ...item,
-        dateTime: formatDate(item.timestamp)
+        dateTime: formatDate(item.timestamp),
+        isWatched: item.statuses.indexOf(ITEM_STATUS.WATCHED) > -1,
+        isNotFound: item.statuses.indexOf(ITEM_STATUS.NOT_FOUND) > -1,
+        hasDecreased: item.statuses.indexOf(ITEM_STATUS.DECREASED) > -1,
+        hasIncreased: item.statuses.indexOf(ITEM_STATUS.INCREASED) > -1,
     }));
     trackedItemsContainer.innerHTML = listItemsLoader({trackedItems});
 
@@ -28,7 +40,7 @@ function addRemoveEvents(selection) {
     for (let elem in deleteElements) {
         if (deleteElements.hasOwnProperty(elem)) {
             const element = deleteElements[elem];
-            const url = element.getAttribute('data-item-url');
+            const url = element.getAttribute("data-item-url");
             element.onclick = removeItem.bind(null, url, element.parentElement);
         }
     }